requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aiohttp==3.6.2            # via discord.py
async-timeout==3.0.1      # via aiohttp
attrs==20.1.0             # via aiohttp
beautifulsoup4==4.9.1     # via bs4
bs4==0.0.1                # via -r requirements.in, pydictionary
certifi==2020.6.20        # via matplotlib, requests
chardet==3.0.4            # via aiohttp, requests
click==7.1.2              # via flask, pydictionary
cycler==0.10.0            # via matplotlib
discord.py==1.4.1         # via -r requirements.in
dnspython==2.0.0          # via -r requirements.in
flask==1.1.2              # via -r requirements.in
futures==3.1.1            # via goslate
fuzzywuzzy==0.18.0        # via -r requirements.in
goslate==1.5.1            # via pydictionary
html2text==2020.1.16      # via -r requirements.in
idna==2.10                # via requests, yarl
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
kiwisolver==1.2.0         # via matplotlib
markupsafe==1.1.1         # via jinja2
matplotlib==3.3.1         # via -r requirements.in
multidict==4.7.6          # via aiohttp, yarl
numpy==1.19.1             # via -r requirements.in, matplotlib
pillow==7.2.0             # via -r requirements.in, matplotlib
pybooru==4.2.0            # via -r requirements.in
pydictionary==2.0.1       # via -r requirements.in
pymongo==3.11.0           # via -r requirements.in
pyparsing==2.4.7          # via matplotlib
python-dateutil==2.8.1    # via matplotlib
python-dotenv==0.14.0     # via -r requirements.in
pytz==2020.1              # via -r requirements.in
requests==2.24.0          # via -r requirements.in, pybooru, pydictionary
six==1.15.0               # via cycler, python-dateutil
soupsieve==2.0.1          # via beautifulsoup4
unidecode==1.1.1          # via -r requirements.in
urllib3==1.25.10          # via requests
werkzeug==1.0.1           # via flask
yarl==1.5.1               # via aiohttp
